requirements.txt
Flask
python-dotenv
openai
pymongo
jwt
Flask-Bcrypt